Create _rules and _parsers files and put them all in the parsing namespace and in the io/parsing folder.
