+++ core/include/bertini2/io/parsing/qi_files.hpp
@@ -1,0 +1,53 @@
+//This file is part of Bertini 2.
+//
+//bertini2/nag_algorithms/zero_dim_solve.hpp is free software: you can redistribute it and/or modify
+//it under the terms of the GNU General Public License as published by
+//the Free Software Foundation, either version 3 of the License, or
+//(at your option) any later version.
+//
+//bertini2/nag_algorithms/zero_dim_solve.hpp is distributed in the hope that it will be useful,
+//but WITHOUT ANY WARRANTY; without even the implied warranty of
+//MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//GNU General Public License for more details.
+//
+//You should have received a copy of the GNU General Public License
+//along with bertini2/nag_algorithms/zero_dim_solve.hpp.  If not, see <http://www.gnu.org/licenses/>.
+//
+// Copyright(C) 2015, 2016 by Bertini2 Development Team
+//
+// See <http://www.gnu.org/licenses/> for a copy of the license,
+// as well as COPYING.  Bertini2 is provided with permitted
+// additional terms in the b2/licenses/ directory.
+
+
+/**
+ \file bertini2/io/parsing/qi_files.hpp
+ 
+ \brief Provides all the include files needed to develop a parsing file that uses boost qi.
+ */
+
+#pragma once
+
+#define BOOST_SPIRIT_USE_PHOENIX_V3 1
+
+#include <boost/fusion/adapted.hpp>
+#include <boost/fusion/include/adapted.hpp>
+
+#include <boost/spirit/include/qi.hpp>
+#include <boost/spirit/include/phoenix.hpp>
+
+
+#include <boost/spirit/include/qi.hpp>
+#include <boost/spirit/include/qi_core.hpp>
+#include <boost/spirit/include/phoenix_core.hpp>
+#include <boost/spirit/include/phoenix_operator.hpp>
+#include <boost/fusion/include/std_pair.hpp>
+
+#include <boost/phoenix/bind/bind_function.hpp>
+#include <boost/phoenix/object/construct.hpp>
+#include <boost/bind.hpp>
+
+#include <boost/fusion/adapted/adt/adapt_adt.hpp>
+#include <boost/fusion/include/adapt_adt.hpp>
+
+#include <boost/spirit/include/support_istream_iterator.hpp>
