+++ core/include/bertini2/io/parsing/settings_parsers.hpp
@@ -1,0 +1,13 @@
+//
+//  settings_parsers.hpp
+//  Xcode_b2
+//
+//  Created by Collins, James B. on 8/19/16.
+//  Copyright (c) 2016 West Texas A&M University. All rights reserved.
+//
+
+#ifndef Xcode_b2_settings_parsers_hpp
+#define Xcode_b2_settings_parsers_hpp
+
+
+#endif
