+++ core/include/bertini2/io/parsing/settings_rules.h
@@ -1,0 +1,13 @@
+//
+//  settings_rules.h
+//  Xcode_b2
+//
+//  Created by Collins, James B. on 8/19/16.
+//  Copyright (c) 2016 West Texas A&M University. All rights reserved.
+//
+
+#ifndef Xcode_b2_settings_rules_h
+#define Xcode_b2_settings_rules_h
+
+
+#endif
